final changes

+++ Assignment-1/include/Mystr.h
@@ -1,8 +1,8 @@
-#include<string.h>
+
 #ifndef __MYSTRING_H
 #define __MYSTRING_H
 
-size_t mystrlen(const char *str);
+int mystrlen(const char *str);
 char* mystrcpy(char* str1, const char* str2);
 char* mystrcat(char *str1, const char *str2);
 int mystrcmp(const char *str1, const char *str2 );
